Update validators db during epoch boundary
