+++ beacon-chain/blockchain/receive_block.go
@@ -6,6 +6,9 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/go-ssz"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
+	"github.com/prysmaticlabs/prysm/beacon-chain/core/validators"
+	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/sirupsen/logrus"
@@ -95,6 +98,13 @@ func (c *ChainService) ReceiveBlockNoPubsub(ctx context.Context, block *ethpb.Be
 		return errors.Wrap(err, "could not clean up block deposits, attestations, and other operations")
 	}
 
+	// Save newly active and deleted validator indices in DB.
+	if helpers.IsEpochStart(block.Slot) {
+		if err := c.updateValidatorsDB(ctx, root); err != nil {
+			return errors.Wrap(err, "could not update validators db")
+		}
+	}
+
 	processedBlkNoPubsub.Inc()
 	return nil
 }
@@ -127,6 +137,13 @@ func (c *ChainService) ReceiveBlockNoPubsubForkchoice(ctx context.Context, block
 		return errors.Wrap(err, "could not clean up block deposits, attestations, and other operations")
 	}
 
+	// Save newly active and deleted validator indices in DB.
+	if helpers.IsEpochStart(block.Slot) {
+		if err := c.updateValidatorsDB(ctx, root); err != nil {
+			return errors.Wrap(err, "could not update validators db")
+		}
+	}
+
 	processedBlkNoPubsubForkchoice.Inc()
 	return nil
 }
@@ -145,5 +162,62 @@ func (c *ChainService) CleanupBlockOperations(ctx context.Context, block *ethpb.
 	for _, dep := range block.Body.Deposits {
 		c.depositCache.RemovePendingDeposit(ctx, dep)
 	}
+	return nil
+}
+
+// this updates validator's pubkey to indices mapping stored in DB, due to validator activation
+// and exit, we should check on every epoch.
+func (c *ChainService) updateValidatorsDB(ctx context.Context, r [32]byte) error {
+	s, err := c.beaconDB.State(ctx, r)
+	if err != nil {
+		return errors.Wrap(err, "could not retrieve latest processed state in DB")
+	}
+	if err := c.saveGenesisValidators(ctx, s); err != nil {
+		return errors.Wrap(err, "could not save validator index")
+	}
+	if err := c.deleteValidatorIdx(ctx, s); err != nil {
+		return errors.Wrap(err, "could not delete validator index")
+	}
+	return nil
+}
+
+// saveValidatorIdx saves the validators public key to index mapping in DB, these
+// validators were activated from current epoch. After it saves, current epoch key
+// is deleted from ActivatedValidators mapping.
+func (c *ChainService) saveValidatorIdx(ctx context.Context, state *pb.BeaconState) error {
+	nextEpoch := helpers.CurrentEpoch(state) + 1
+	activatedValidators := validators.ActivatedValFromEpoch(nextEpoch)
+	var idxNotInState []uint64
+	for _, idx := range activatedValidators {
+		// If for some reason the activated validator indices is not in state,
+		// we skip them and save them to process for next epoch.
+		if int(idx) >= len(state.Validators) {
+			idxNotInState = append(idxNotInState, idx)
+			continue
+		}
+		pubKey := state.Validators[idx].PublicKey
+		if err := c.beaconDB.SaveValidatorIndex(ctx, bytesutil.ToBytes48(pubKey), idx); err != nil {
+			return errors.Wrap(err, "could not save validator index")
+		}
+	}
+	// Since we are processing next epoch, save the can't processed validator indices
+	// to the epoch after that.
+	validators.InsertActivatedIndices(nextEpoch+1, idxNotInState)
+	validators.DeleteActivatedVal(helpers.CurrentEpoch(state))
+	return nil
+}
+
+// deleteValidatorIdx deletes the validators public key to index mapping in DB, the
+// validators were exited from current epoch. After it deletes, current epoch key
+// is deleted from ExitedValidators mapping.
+func (c *ChainService) deleteValidatorIdx(ctx context.Context, state *pb.BeaconState) error {
+	exitedValidators := validators.ExitedValFromEpoch(helpers.CurrentEpoch(state) + 1)
+	for _, idx := range exitedValidators {
+		pubKey := state.Validators[idx].PublicKey
+		if err := c.beaconDB.DeleteValidatorIndex(ctx, bytesutil.ToBytes48(pubKey)); err != nil {
+			return errors.Wrap(err, "could not delete validator index")
+		}
+	}
+	validators.DeleteExitedVal(helpers.CurrentEpoch(state))
 	return nil
 }
